done all Elena's remarks - first try

+++ src/main/java/core/basesyntax/operationhandlers/PurchaseOperationHandler.java
@@ -6,6 +6,9 @@ import core.basesyntax.model.FruitTransactionImpl;
 public class PurchaseOperationHandler implements OperationHandler {
     @Override
     public void apply(FruitTransactionImpl transaction) {
+        if (transaction == null) {
+            throw new IllegalArgumentException("Transaction cannot be null");
+        }
         Storage.removeFruit(transaction.getFruit(), transaction.getQuantity());
     }
 }
